[stream_orderer] Add flag to indicate end of input
The stream orderer is used in various contexts, and in some—such as an empty
thin-provisioned block device—we hang because we can't detect that input is
complete. This change adds a flag to explicitly signal when the producer is
done feeding items into the stream orderer.

This is a workaround rather than a proper fix; ideally, the code should be
refactored to provide a single, reliable mechanism for detecting completion.

In some scenarios, we need to distinguish between:

    1. All current items in the stream orderer have been processed.
    2. No more items will be added.

This change enables that distinction. I’ll revisit and improve this design if
the functionality is fully integrated.

Signed-off-by: Tony Asleson <tasleson@redhat.com>