Change validate_slab_offsets_file to return tuple
Changed the return type from Result<bool> to (bool, u32, Option<String>)
to provide more detailed error information and the slab count directly.
This was mainly done as we don't what this to return an Err as we are
using this when we are trying to figure out what is still valid.

Signed-off-by: Tony Asleson <tasleson@redhat.com>